Reduce extraneous maodles in Puppetfile
Prior to this commit, there were several modules in the Puppetfile
that aren't used by any of the materials in the README.  They were
originall included to help bootstrap some testing with different
git servers.

After this commit, the Puppetfile is reduced to a set of modules
that are core to the use of the control-repo.