Replace :latest with actual version to follow best practice
Prior to this commit, multiple modules specified a :latest version.
This is poor practice and should never be used in production.  It
was originally this way because I didn't want to update the versions
myself every time a new one came out.

After this commit, we use specific versions for all modules in
alignment with best practice.