Update README to use webhook_url.txt
Prior to this commit, we relied on using abrader-gms to setup the
webhook in a user's git server.

After this commit, we no longer rely on abrader-gms.  The
pe_code_manager_webhook module lays down a file that contains the
webhook url that users should paste into their Git server's UI.
This reduces complexity of the starting steps of using the control
repo and makes documenting multiple different git server UI's easier.