Move puppet master profile back into the control-repo
Prior to this commit, the rampup_profile_puppetmaster was included
via the Puppetfile.  This was all fine and dandy until you wanted
to change your hiera hierarchy and you realize you need to fork
the module and edit a line to get what you want.

After this commit, the puppetmaster profile is included directly
in the control-repo.  This way when you want to edit your hiera
hierarchy you can simply edit the code in your control-repo and
continue on your day.